moved

+++ examples/HRC20.sol
@@ -1,11 +1,11 @@
 //SPDX-License-Identifier: Unlicense
 pragma solidity ^0.8.0;
 
-import "../SimpleTerms.sol";
+import "../contracts/SimpleTerms.sol";
 
 // This is a demonstation of how to use SimpleTerms, note the use of the checkAcceptance modifier!
 
-contract HRC20 is SimpleTerms{
+contract HRC20 is SimpleTerms {
     // Public variables of the token
     string public name;
     string public symbol;
@@ -14,14 +14,18 @@ contract HRC20 is SimpleTerms{
     uint256 public totalSupply;
 
     // This creates an array with all balances
-    mapping (address => uint256) public balanceOf;
-    mapping (address => mapping (address => uint256)) public allowance;
+    mapping(address => uint256) public balanceOf;
+    mapping(address => mapping(address => uint256)) public allowance;
 
     // This generates a public event on the blockchain that will notify clients
     event Transfer(address indexed from, address indexed to, uint256 value);
-    
+
     // This generates a public event on the blockchain that will notify clients
-    event Approval(address indexed _owner, address indexed _spender, uint256 _value);
+    event Approval(
+        address indexed _owner,
+        address indexed _spender,
+        uint256 _value
+    );
 
     // This notifies clients about the amount burnt
     event Burn(address indexed from, uint256 value);
@@ -36,18 +40,22 @@ contract HRC20 is SimpleTerms{
         string memory tokenSymbol,
         uint256 initialSupply,
         uint8 _decimals
-    )  {
+    ) {
         decimals = _decimals;
-        totalSupply = initialSupply * 10 ** uint256(decimals);  // Update total supply with the decimal amount
-        balanceOf[msg.sender] = totalSupply;                // Give the creator all initial tokens
-        name = tokenName;                                   // Set the name for display purposes
-        symbol = tokenSymbol;                               // Set the symbol for display purposes
+        totalSupply = initialSupply * 10**uint256(decimals); // Update total supply with the decimal amount
+        balanceOf[msg.sender] = totalSupply; // Give the creator all initial tokens
+        name = tokenName; // Set the name for display purposes
+        symbol = tokenSymbol; // Set the symbol for display purposes
     }
 
     /**
      * Internal transfer, only can be called by this contract
      */
-    function _transfer(address _from, address _to, uint _value) internal {
+    function _transfer(
+        address _from,
+        address _to,
+        uint256 _value
+    ) internal {
         // Prevent transfer to 0x0 address. Use burn() instead
         require(_to != address(0x0));
         // Check if the sender has enough
@@ -55,7 +63,7 @@ contract HRC20 is SimpleTerms{
         // Check for overflows
         require(balanceOf[_to] + _value >= balanceOf[_to]);
         // Save this for an assertion in the future
-        uint previousBalances = balanceOf[_from] + balanceOf[_to];
+        uint256 previousBalances = balanceOf[_from] + balanceOf[_to];
         // Subtract from the sender
         balanceOf[_from] -= _value;
         // Add the same to the recipient
@@ -73,7 +81,11 @@ contract HRC20 is SimpleTerms{
      * @param _to The address of the recipient
      * @param _value the amount to send
      */
-    function transfer(address _to, uint256 _value) public checkAcceptance returns (bool success){
+    function transfer(address _to, uint256 _value)
+        public
+        checkAcceptance
+        returns (bool success)
+    {
         _transfer(msg.sender, _to, _value);
         return true;
     }
@@ -87,8 +99,12 @@ contract HRC20 is SimpleTerms{
      * @param _to The address of the recipient
      * @param _value the amount to send
      */
-    function transferFrom(address _from, address _to, uint256 _value) public checkAcceptance returns (bool success) {
-        require(_value <= allowance[_from][msg.sender]);     // Check allowance
+    function transferFrom(
+        address _from,
+        address _to,
+        uint256 _value
+    ) public checkAcceptance returns (bool success) {
+        require(_value <= allowance[_from][msg.sender]); // Check allowance
         allowance[_from][msg.sender] -= _value;
         _transfer(_from, _to, _value);
         return true;
@@ -102,8 +118,11 @@ contract HRC20 is SimpleTerms{
      * @param _spender The address authorized to spend
      * @param _value the max amount they can spend
      */
-    function approve(address _spender, uint256 _value) public checkAcceptance
-        returns (bool success) {
+    function approve(address _spender, uint256 _value)
+        public
+        checkAcceptance
+        returns (bool success)
+    {
         allowance[msg.sender][_spender] = _value;
         emit Approval(msg.sender, _spender, _value);
         return true;
